Re-Setup Library Files
Needed to find out why the program wasn't running, so setup the project again and had a headache trying to figure out why SDL.h wasn't recognised. Turns out it had switched to Win32 rather than x64, so had to re-add linker paths etc.
Now there's an issue with not recognising _mixer and _ttf methods, but the problem was a missing DLL file (that's supposed to come with Windows 10 anyway??) so now it works. Yay!